requirements.txt
streamlit==1.24.0
requests==2.28.1
python-dotenv==1.0.0
transformers==4.32.0
langchain==0.0.232
